[PATCH] blkparse: Add per-process stats and fix output
Add the option (-s) to display per-process read/write stats as well as
the per-processor and total output. At the same time, fix the output so
it's a single file instead of per-processor (doesn't make sense here) and
put everything in the output file given with -o. If no output file is given,
stdout is used.