[PATCH] blkparse: more optimizations
The sort path unification and memory cutback introduced some
overhead, mainly due to lots of memory allocations and frees
of the same structures over and over again. Add a frontend
cache for struct trace and blk_io_trace allocations.

Also concatenate a lot of the string operations in fmt_select,
it's quite high in the profiles.

This brings us to the same performance as the memory hog version.