[PATCH] blkparse: sequence fixes
This should finally fix all the remaining sequence bugs. We
keep a backlog of already processed entries so we can lookup
any out-of-sequence events there. That takes care of one
direction of sequence and time mismatches. The other direction
is handled by fixing the trace_rb_find() lookup to work well
enough without having a precise time (which is the primary
key for the rb sort).