[PATCH] blkparse: Remove negative skips
The following patch does not insert skips that go backwards. I am not
sure why we do, but on occasion we find events that are behind where we
should be, and we don't have an entry in skipped ones - that would make
me thing it's a duplicate. But in any event, we should not be reporting
such things - they cause huge skip numbers by wrapping on an unsigned long.