[PATCH] blkparse: don't do partial reads in read_data()
If we do a partial read, the trace may look ok because the magic is there,
but we will have lost it and the next read() will be confused. So only
return early if we read zero bytes and we must not block, otherwise wait
around for the remaining data.