Changed triggering on INSERT to GETRQ (temporary?)
This is backwards compatible with previous versions, with the one
difference being that the Q2I information may be a bit shorter.

The changes forthcoming with the explicit unplugging (2.6.20 + Jens'
unplag patches) require this. We may need to redefine and/or add a new
measurement once that settles down. When that happens, we'll most likely be
renaming things to better reflect GETRQ versus INSERT.

Signed-off-by: Alan D. Brunelle <Alan.Brunelle@hp.com
Signed-off-by: Jens Axboe <jens.axboe@oracle.com>