Added active requests at Q information.
An important consideration when analyzing block IO schedulers is to know
how many requests the scheduler has to work with. The metric provided
in this section details how many requests (on average) were being held
by the IO scheduler when an incoming IO request was being handled. To
determine this, btt keeps track of how many Q requests came in, and
subtacts requests that have been issued (D).

Sample:

==================== Active Requests At Q Information ====================

       DEV |  Avg Reqs @ Q
---------- | -------------
 ( 65, 80) |          12.0
 ( 65,240) |          16.9
 ( 65,112) |          13.3
 ( 66, 64) |          32.8
 ( 66, 80) |          21.5
 ( 65, 96) |           8.6
 ( 66, 16) |          16.2
 ( 65, 64) |          20.4
 ( 66, 96) |          18.3
 ( 65,176) |          17.4
 ( 66, 32) |          13.6
 ( 65,144) |          13.4
 ( 66,  0) |          21.4
 ( 65,192) |          19.4
 ( 66,128) |          18.4
 ( 66,144) |          16.2
 ( 65,128) |           8.0
 ( 66,112) |          44.2
---------- | -------------
   Overall | Avgs Reqs @ Q
   Average |          17.4

Signed-off-by: Alan D. Brunelle <Alan.Brunelle@hp.com>