Fix Q counts during requeue and merges.
It looks to me like btt doesn't correctly keep track
of the number of requests currently in the queue for
a device.  n_act_q gets incremented in trace_queue
and decremented in trace_issue, but I think it also
needs to get updated in trace_merge and trace_requeue.

The one thing I'm not sure about is whether we want
r_iop->dip->n_qs++ in the new handle_requeue routine.

The following patch makes the "active requests at Q"
count a little more sane for me.  This is against
git as of yesterday.
Signed-off-by: Alan D. Brunelle <Alan.Brunelle@hp.com>