blkparse: Introduce optional accounting of PC requests
The current blkparse lists only PC requests, but does not account for
number of PC requests and the size of the attached data.  This patch
introduces optional accounting for PC requests.  If any PC event was
encountered, the summary output will also include data from PC
requests.

For PC completion events the amount of transferred data is not
available, since the kernel uses the data_len field as residual bytes
count on event completion.

Signed-off-by: Christof Schmitt <christof.schmitt@de.ibm.com>
Signed-off-by: Jens Axboe <jens.axboe@oracle.com>