btt: Added no remap option
Trying to run btt on pre-2.6.19 kernels has problems handling the
previous remap PDU - it did not include a proper device-from field.
(Probably should have bumped the blktrace version when we did that.)

This option just tosses those out as it just results in lots of crazy
stuff being handled.

Signed-off-by: Alan D. Brunelle <alan.brunelle@hp.com>