fix max-pkts option inconsistencies
This is for RH bug 498426,
btrecord doesn't recognize --max-pkts and --max-packets cmd line options

Usage text and some documentation references "--max-pkts" and
the man page references "--max-packets" but the getopts code
is looking for --max_pkts.

I'm not sure if it's best to make the code match the majority
of the docs, or fix the docs to match the code?  This patch does
the former, but I'm not picky if you want to go the other way :)

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Jens Axboe <jens.axboe@oracle.com>