btt: Added in I/O activity per device and system-wide
It now keeps track of I/O activity on a per-device basis (as well as a
cumulative system-wide view). ``I/O activity'' is defined as defined as
the time during which the device driver and device are activelty working
on at least one I/O. Here's a sample output:

==================== I/O Active Period Information ====================

       DEV |     # Live      Avg. Act     Avg. !Act % Live
---------- | ---------- ------------- ------------- ------
 (254,  0) |          0   0.000000000   0.000000000   0.00
 (  8, 17) |          0   0.000000000   0.000000000   0.00
 (  8, 16) |         29   0.909596815   0.094646263  90.87
 (  8, 33) |          0   0.000000000   0.000000000   0.00
 (  8, 32) |        168   0.097848226   0.068231948  59.06
---------- | ---------- ------------- ------------- ------
 Total Sys |         33   0.799808811   0.082334758  90.92

Also added a new btt -Z option that generates per-device and system-wide
I/O activity data that can be plotted.

Refer to the documentation updates (btt.1, btt.tex) for more information.