iowatcher: Fix processing of trace filenames containing spaces
blktrace_to_dump passes filenames containing spaces to blkparse via
system() so only the first chunk of the string is taken to be the
filename by the subprocess.

This switches to using posix_spawnp() so that we can present the
filename as an element of argv and avoid iowatcher failing in these
cases.

Signed-off-by: Andrew Price <anprice@redhat.com>